feat(home): add view varaibles config
